refactor moderation reports and labeler settings
