+++ lib/src/core/network/atproto/data/repositories/repo_repository.dart
@@ -73,12 +73,10 @@ abstract class RepoRepository {
   /// [subject] The subject of the report (content or account)
   /// [reasonType] The reason for the report
   /// [reason] Optional additional context about the violation
-  /// [service] Optional moderation service to use
+  /// [serviceDid] Moderation service receiving the report
   ///
-  /// Returns true if the report was successfully created
-  Future<bool> createReport({
+  Future<void> createReport({
     required ModerationCreateReportInput input,
-    dynamic service,
-    String? serviceDid,
+    required String serviceDid,
   });
 }
