+++ lib/src/core/ui/widgets/report_dialog.dart
@@ -1,54 +1,16 @@
 import 'dart:async';
 
 import 'package:poptart_lex/com/atproto/moderation/create_report.dart';
-import 'package:poptart_lex/com/atproto/moderation/defs.dart';
 import 'package:auto_route/auto_route.dart';
-import 'package:poptart_lex/com/atproto/repo/strong_ref.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
 import 'package:spark/src/core/design_system/components/atoms/buttons/app_button.dart';
 import 'package:spark/src/core/l10n/app_localizations.dart';
-import 'package:spark/src/core/network/atproto/data/repositories/labeler_repository.dart';
-import 'package:spark/src/core/network/atproto/data/repositories/sprk_repository.dart';
+import 'package:spark/src/core/moderation/moderation_report_catalog.dart';
+import 'package:spark/src/core/moderation/moderation_report_service.dart';
 import 'package:spark/src/core/utils/logging/log_service.dart';
 import 'package:spark/src/core/utils/logging/logger.dart';
-
-typedef ReportSubmitCallback =
-    Future<void> Function(
-      UModerationCreateReportSubject subject,
-      ReasonType reasonType,
-      String? reason,
-      String serviceDid,
-    );
-
-enum ReportCategory {
-  violence('Violence'),
-  sexual('Sexual Content'),
-  childSafety('Child Safety'),
-  harassment('Harassment'),
-  misleading('Misleading'),
-  ruleViolations('Rule Violations'),
-  selfHarm('Self-Harm'),
-  other('Other');
-
-  const ReportCategory(this.displayName);
-  final String displayName;
-}
-
-class ReportReason {
-  final String value;
-  final String displayName;
-  final String? description;
-  final KnownReasonType? knownType;
-
-  const ReportReason({
-    required this.value,
-    required this.displayName,
-    this.description,
-    this.knownType,
-  });
-}
 
 class ReportDialog extends ConsumerStatefulWidget {
   const ReportDialog({
@@ -60,9 +22,7 @@ class ReportDialog extends ConsumerStatefulWidget {
   final UModerationCreateReportSubject subject;
   final String? fallbackServiceDid;
 
-  /// Callback for report submission. Uses [ReasonType] directly to support
-  /// known & unknown reason types.
-  final ReportSubmitCallback? onSubmit;
+  final ModerationReportSubmitter? onSubmit;
 
   @override
   ConsumerState<ReportDialog> createState() => _ReportDialogState();
@@ -78,271 +38,8 @@ class _ReportDialogState extends ConsumerState<ReportDialog> {
       TextEditingController();
   bool _isSubmitting = false;
   String? _errorMessage;
-  Future<List<CompatibleModerationService>>? _compatibleServicesFuture;
+  Future<ModerationReportOptions>? _reportOptionsFuture;
   String? _selectedServiceDid;
-
-  // Map categories to their reasons
-  static final Map<ReportCategory, List<ReportReason>> _categoryReasons = {
-    ReportCategory.violence: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonViolenceAnimal',
-        displayName: 'Animal Abuse',
-        description: 'Content depicting harm to animals',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonViolenceAnimal,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonViolenceThreats',
-        displayName: 'Threats',
-        description: 'Threats of violence',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonViolenceThreats,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonViolenceGraphicContent',
-        displayName: 'Graphic Content',
-        description: 'Graphic or violent imagery',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonViolenceGraphicContent,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonViolenceGlorification',
-        displayName: 'Glorification of Violence',
-        description: 'Content that glorifies violence',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonViolenceGlorification,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonViolenceExtremistContent',
-        displayName: 'Extremist Content',
-        description: 'Content promoting extremist ideologies',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonViolenceExtremistContent,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonViolenceTrafficking',
-        displayName: 'Trafficking',
-        description: 'Content related to human trafficking',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonViolenceTrafficking,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonViolenceOther',
-        displayName: 'Other Violence',
-        description: 'Other violent content',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonViolenceOther,
-      ),
-    ],
-    ReportCategory.sexual: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSexualAbuseContent',
-        displayName: 'Abuse Content',
-        description: 'Sexual abuse content',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSexualAbuseContent,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSexualNCII',
-        displayName: 'Non-Consensual Intimate Images',
-        description: 'Sharing intimate images without consent',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSexualNCII,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSexualDeepfake',
-        displayName: 'Deepfake',
-        description: 'AI-generated sexual content',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSexualDeepfake,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSexualAnimal',
-        displayName: 'Animal Sexual Content',
-        description: 'Sexual content involving animals',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSexualAnimal,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSexualUnlabeled',
-        displayName: 'Unlabeled Sexual Content',
-        description: 'Sexual content without proper warnings',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSexualUnlabeled,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSexualOther',
-        displayName: 'Other Sexual Content',
-        description: 'Other sexual content violations',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSexualOther,
-      ),
-    ],
-    ReportCategory.childSafety: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonChildSafetyCSAM',
-        displayName: 'CSAM',
-        description: 'Child sexual abuse material',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonChildSafetyCSAM,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonChildSafetyGroom',
-        displayName: 'Grooming',
-        description: 'Grooming behavior targeting minors',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonChildSafetyGroom,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonChildSafetyPrivacy',
-        displayName: 'Privacy Violation',
-        description: 'Sharing private information about minors',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonChildSafetyPrivacy,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonChildSafetyHarassment',
-        displayName: 'Harassment',
-        description: 'Harassment targeting minors',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonChildSafetyHarassment,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonChildSafetyOther',
-        displayName: 'Other Child Safety',
-        description: 'Other child safety concerns',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonChildSafetyOther,
-      ),
-    ],
-    ReportCategory.harassment: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonHarassmentTroll',
-        displayName: 'Trolling',
-        description: 'Trolling or disruptive behavior',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonHarassmentTroll,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonHarassmentTargeted',
-        displayName: 'Targeted Harassment',
-        description: 'Targeted harassment or bullying',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonHarassmentTargeted,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonHarassmentHateSpeech',
-        displayName: 'Hate Speech',
-        description: 'Hate speech or discriminatory content',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonHarassmentHateSpeech,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonHarassmentDoxxing',
-        displayName: 'Doxxing',
-        description: 'Sharing private information without consent',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonHarassmentDoxxing,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonHarassmentOther',
-        displayName: 'Other Harassment',
-        description: 'Other harassment violations',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonHarassmentOther,
-      ),
-    ],
-    ReportCategory.misleading: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonMisleadingBot',
-        displayName: 'Bot Account',
-        description: 'Automated or bot account',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonMisleadingBot,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonMisleadingImpersonation',
-        displayName: 'Impersonation',
-        description: 'Impersonating another person or entity',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonMisleadingImpersonation,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonMisleadingSpam',
-        displayName: 'Spam',
-        description: 'Spam or repetitive content',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonMisleadingSpam,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonMisleadingScam',
-        displayName: 'Scam',
-        description: 'Fraudulent or scam content',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonMisleadingScam,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonMisleadingElections',
-        displayName: 'Election Misinformation',
-        description: 'False information about elections',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonMisleadingElections,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonMisleadingOther',
-        displayName: 'Other Misleading',
-        description: 'Other misleading content',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonMisleadingOther,
-      ),
-    ],
-    ReportCategory.ruleViolations: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonRuleSiteSecurity',
-        displayName: 'Site Security',
-        description: 'Violation of site security rules',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonRuleSiteSecurity,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonRuleProhibitedSales',
-        displayName: 'Prohibited Sales',
-        description: 'Prohibited goods or services',
-        knownType:
-            KnownReasonType.toolsOzoneReportDefsReasonRuleProhibitedSales,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonRuleBanEvasion',
-        displayName: 'Ban Evasion',
-        description: 'Attempting to evade a ban',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonRuleBanEvasion,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonRuleOther',
-        displayName: 'Other Rule Violation',
-        description: 'Other rule violations',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonRuleOther,
-      ),
-    ],
-    ReportCategory.selfHarm: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSelfHarmContent',
-        displayName: 'Self-Harm Content',
-        description: 'Content promoting self-harm',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSelfHarmContent,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSelfHarmED',
-        displayName: 'Eating Disorder',
-        description: 'Content promoting eating disorders',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSelfHarmED,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSelfHarmStunts',
-        displayName: 'Dangerous Stunts',
-        description: 'Content showing dangerous stunts',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSelfHarmStunts,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSelfHarmSubstances',
-        displayName: 'Substance Abuse',
-        description: 'Content promoting substance abuse',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSelfHarmSubstances,
-      ),
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonSelfHarmOther',
-        displayName: 'Other Self-Harm',
-        description: 'Other self-harm related content',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonSelfHarmOther,
-      ),
-    ],
-    ReportCategory.other: [
-      const ReportReason(
-        value: 'tools.ozone.report.defs#reasonOther',
-        displayName: 'Other',
-        description: 'Other issues not listed above',
-        knownType: KnownReasonType.toolsOzoneReportDefsReasonOther,
-      ),
-    ],
-  };
 
   @override
   void dispose() {
@@ -358,24 +55,30 @@ class _ReportDialogState extends ConsumerState<ReportDialog> {
   }
 
   void _selectReason(ReportReason reason) {
-    final servicesFuture = _loadCompatibleServices(reason);
+    final optionsFuture = ref
+        .read(moderationReportServiceProvider)
+        .loadOptions(
+          subject: widget.subject,
+          reasonType: reason.reasonType,
+          fallbackServiceDid: widget.fallbackServiceDid,
+        );
     setState(() {
       _selectedReason = reason;
-      _compatibleServicesFuture = servicesFuture;
+      _reportOptionsFuture = optionsFuture;
       _selectedServiceDid = null;
     });
-    unawaited(_selectDefaultService(servicesFuture, reason));
+    unawaited(_selectDefaultService(optionsFuture, reason));
   }
 
   Future<void> _selectDefaultService(
-    Future<List<CompatibleModerationService>> servicesFuture,
+    Future<ModerationReportOptions> optionsFuture,
     ReportReason reason,
   ) async {
     try {
-      final services = await servicesFuture;
-      if (!mounted || _selectedReason != reason || services.isEmpty) return;
+      final options = await optionsFuture;
+      if (!mounted || _selectedReason != reason) return;
       setState(() {
-        _selectedServiceDid = services.first.did;
+        _selectedServiceDid = options.defaultServiceDid;
       });
     } catch (error, stackTrace) {
       _logger.w(
@@ -390,37 +93,9 @@ class _ReportDialogState extends ConsumerState<ReportDialog> {
     setState(() {
       _selectedCategory = null;
       _selectedReason = null;
-      _compatibleServicesFuture = null;
+      _reportOptionsFuture = null;
       _selectedServiceDid = null;
     });
-  }
-
-  ReasonType _reasonTypeFor(ReportReason reason) {
-    return reason.knownType != null
-        ? ReasonType.knownValue(data: reason.knownType!)
-        : ReasonType.unknown(data: reason.value);
-  }
-
-  Future<List<CompatibleModerationService>> _loadCompatibleServices(
-    ReportReason reason,
-  ) async {
-    final repository = GetIt.instance<SprkRepository>();
-    final subject = widget.subject.data;
-    final isBskyRecord =
-        subject is RepoStrongRef &&
-        subject.uri.collection.toString().startsWith('app.bsky');
-    final fallbackProxyDid =
-        widget.fallbackServiceDid ??
-        (isBskyRecord ? repository.bskyModDid : repository.modDid);
-    final fallbackDid = fallbackProxyDid.split('#').first;
-    return repository.labeler.getCompatibleModerationServices(
-      repository.labelerDids,
-      ModerationServiceQuery.forReport(
-        fallbackDid: fallbackDid,
-        subject: widget.subject,
-        reasonType: _reasonTypeFor(reason).toJson(),
-      ),
-    );
   }
 
   Future<void> _submitReport() async {
@@ -438,37 +113,17 @@ class _ReportDialogState extends ConsumerState<ReportDialog> {
     });
 
     try {
-      // Build the ReasonType: use known type if available, otherwise unknown
-      // with the raw value
-      final reasonType = _reasonTypeFor(_selectedReason!);
-
-      if (widget.onSubmit != null) {
-        await widget.onSubmit!(
-          subject,
-          reasonType,
-          reason,
-          _selectedServiceDid!,
-        );
-        if (mounted) {
-          context.router.maybePop();
-        }
-      } else {
-        // Get the repository directly and create the report
-        final repoRepository = GetIt.instance<SprkRepository>().repo;
-        _logger.d('Creating report with reason: ${_selectedReason!.value}');
-
-        final success = await repoRepository.createReport(
-          input: ModerationCreateReportInput(
+      await ref
+          .read(moderationReportServiceProvider)
+          .submit(
             subject: subject,
-            reasonType: reasonType,
+            reasonType: _selectedReason!.reasonType,
             reason: reason,
-          ),
-          serviceDid: _selectedServiceDid,
-        );
-
-        if (success && mounted) {
-          context.router.maybePop();
-        }
+            serviceDid: _selectedServiceDid!,
+            submitter: widget.onSubmit,
+          );
+      if (mounted) {
+        context.router.maybePop();
       }
     } catch (e) {
       _logger.e('Error creating report', error: e);
@@ -495,7 +150,7 @@ class _ReportDialogState extends ConsumerState<ReportDialog> {
         (theme.brightness == Brightness.dark ? Colors.white : Colors.black);
     final isStep2 = _selectedCategory != null;
     final reasons = isStep2
-        ? (_categoryReasons[_selectedCategory!] ?? <ReportReason>[])
+        ? (reportCategoryReasons[_selectedCategory!] ?? <ReportReason>[])
         : <ReportReason>[];
 
     return AlertDialog(
@@ -564,8 +219,8 @@ class _ReportDialogState extends ConsumerState<ReportDialog> {
 
               if (isStep2 && _selectedReason != null) ...[
                 const SizedBox(height: 8),
-                FutureBuilder<List<CompatibleModerationService>>(
-                  future: _compatibleServicesFuture,
+                FutureBuilder<ModerationReportOptions>(
+                  future: _reportOptionsFuture,
                   builder: (context, snapshot) {
                     if (snapshot.connectionState != ConnectionState.done) {
                       return const Center(
@@ -578,7 +233,7 @@ class _ReportDialogState extends ConsumerState<ReportDialog> {
                     if (snapshot.hasError) {
                       return Text(l10n.moderationServiceLoadFailed);
                     }
-                    final services = snapshot.data ?? const [];
+                    final services = snapshot.data?.services ?? const [];
                     if (services.isEmpty) {
                       return Text(l10n.moderationNoCompatibleService);
                     }
